Remove SSLWarningPanel and NextGen ad
A similar message was present in both a tooltip on hover and a popup
panel on left-click. This commit removes the SSLWarningPanel and relies
on the tooltip for messaging.

The message has been changed to indicate a compatible TLS plugin was not
detected without including a sales pitch. Additionally it now says that
TLS security options and mTLS are not configurable through the interface
rather than saying they are not supported.

The icon for when the message is present has been changed from a lock
with a red X to an information blue circle with an "i" to indicate this
is not an error or security issue.

Issue: https://github.com/OpenIntegrationEngine/engine/issues/156
Co-authored-by: Tony Germano <tony@germano.name>
Signed-off-by: Jon Bartels <jon.bartels@teladochealth.com>
Signed-off-by: Tony Germano <tony@germano.name>